re organize code
