+++ src/components/CurrencySelectorElement.tsx
@@ -6,10 +6,93 @@ import {
   StyleSheet,
 } from 'react-native';
 import type { Checkout } from '../types/Checkout';
-import type { CurrencySelectorAppearance } from '../types/PaymentSheet';
+import type { FontConfig, ThemedColor } from '../types/PaymentSheet';
 import NativeCurrencySelectorElement, {
   type HeightChangeEvent,
 } from '../specs/NativeCurrencySelectorElement';
+
+/**
+ * Controls what content is displayed in each currency option's label.
+ */
+export enum CurrencySelectorLabelContent {
+  /**
+   * Automatically determines the best display based on the purchase type.
+   * For one-time payments this shows the formatted amount; for recurring
+   * payments this shows the currency code.
+   */
+  Automatic = 'automatic',
+
+  /** Displays only the currency code (e.g. "USD"). */
+  CurrencyCode = 'currencyCode',
+
+  /** Displays the formatted amount (e.g. "$12.00"). */
+  Amount = 'amount',
+}
+
+/**
+ * Appearance configuration for CurrencySelectorElement.
+ *
+ * All properties are optional — when omitted, the element uses
+ * platform-appropriate defaults that adapt to light/dark mode.
+ */
+export interface CurrencySelectorAppearance {
+  /** Background color of the selector track. */
+  background?: ThemedColor;
+
+  /** Background color of the selected currency pill. */
+  selectedBackground?: ThemedColor;
+
+  /** Border color for the track outline. */
+  borderColor?: ThemedColor;
+
+  /** Text color for unselected currency options. */
+  textColor?: ThemedColor;
+
+  /** Text color for the currently selected currency option. */
+  selectedTextColor?: ThemedColor;
+
+  /**
+   * Text color for the exchange rate caption and disclosure text.
+   *
+   * Note: Alpha values below 0.5 are clamped to 0.5 on both platforms
+   * to maintain regulatory text legibility.
+   */
+  textSecondaryColor?: ThemedColor;
+
+  /** Color for error messages shown below the selector. */
+  dangerColor?: ThemedColor;
+
+  /**
+   * Corner radius applied to the track and selected currency pill, in
+   * density-independent pixels.
+   *
+   * When omitted, the element uses a full capsule shape (50% rounding).
+   */
+  cornerRadius?: number;
+
+  /** Border width for the track outline, in density-independent pixels. */
+  borderWidth?: number;
+
+  /**
+   * Vertical padding between the track edges and the currency option
+   * content, in density-independent pixels.
+   */
+  contentVerticalPadding?: number;
+
+  /**
+   * Font configuration for the selector labels and captions.
+   *
+   * Uses the same `FontConfig` shape as PaymentSheet.
+   */
+  font?: Partial<FontConfig>;
+
+  /**
+   * Controls what is displayed in each currency option's label.
+   *
+   * @default CurrencySelectorLabelContent.Automatic
+   */
+  labelContent?: CurrencySelectorLabelContent;
+}
 
 /**
  * Props for {@link CurrencySelectorElement}.
