+++ src/types/PaymentSheet.ts
@@ -432,144 +432,6 @@ export type PrimaryButtonColorConfig = {
 /** A color that's either a single hex or a light/dark pair */
 export type ThemedColor = string | { light: string; dark: string };
 
-/**
- * Controls what content is displayed in each currency option's label.
- */
-export enum CurrencySelectorLabelContent {
-  /**
-   * Automatically determines the best display based on the purchase type.
-   * For one-time payments this shows the formatted amount; for recurring
-   * payments this shows the currency code.
-   */
-  Automatic = 'automatic',
-
-  /** Displays only the currency code (e.g. "USD"). */
-  CurrencyCode = 'currencyCode',
-
-  /** Displays the formatted amount (e.g. "$12.00"). */
-  Amount = 'amount',
-}
-
-/**
- * Appearance configuration for CurrencySelectorElement.
- *
- * All properties are optional — when omitted, the element uses
- * platform-appropriate defaults that adapt to light/dark mode.
- */
-export interface CurrencySelectorAppearance {
-  // ─── Colors ──────────────────────────────────────────────────────────
-
-  /**
-   * Background color of the selector track.
-   *
-   * @default iOS: UIColor.tertiarySystemFill
-   * @default Android: Theme surface color
-   */
-  background?: ThemedColor;
-
-  /**
-   * Background color of the selected currency pill.
-   *
-   * @default iOS: white (light) / tertiaryLabel (dark)
-   * @default Android: Theme primary color
-   */
-  selectedBackground?: ThemedColor;
-
-  /**
-   * Border color for the track outline.
-   *
-   * @default iOS: UIColor.separator
-   * @default Android: Theme onSurface at 12% alpha
-   */
-  borderColor?: ThemedColor;
-
-  /**
-   * Text color for unselected currency options.
-   *
-   * @default iOS: UIColor.label
-   * @default Android: Theme onComponent color
-   */
-  textColor?: ThemedColor;
-
-  /**
-   * Text color for the currently selected currency option.
-   *
-   * @default iOS: UIColor.label
-   * @default Android: Theme onPrimary color
-   */
-  selectedTextColor?: ThemedColor;
-
-  /**
-   * Text color for the exchange rate caption and disclosure text.
-   *
-   * Note: Alpha values below 0.5 are clamped to 0.5 on both platforms
-   * to maintain regulatory text legibility.
-   *
-   * @default iOS: UIColor.secondaryLabel
-   * @default Android: Theme subtitle color
-   */
-  textSecondaryColor?: ThemedColor;
-
-  /**
-   * Color for error messages shown below the selector.
-   *
-   * @default iOS: UIColor.systemRed
-   * @default Android: Theme error color
-   */
-  dangerColor?: ThemedColor;
-
-  // ─── Dimensions ──────────────────────────────────────────────────────
-
-  /**
-   * Corner radius applied to the track and selected currency pill, in
-   * density-independent pixels.
-   *
-   * When omitted, the element uses a full capsule shape (50% rounding).
-   *
-   * @default undefined (capsule)
-   */
-  cornerRadius?: number;
-
-  /**
-   * Border width for the track outline, in density-independent pixels.
-   *
-   * @default iOS: 0
-   * @default Android: 1
-   */
-  borderWidth?: number;
-
-  /**
-   * Vertical padding between the track edges and the currency option
-   * content, in density-independent pixels.
-   *
-   * @default iOS: 3
-   * @default Android: 4
-   */
-  contentVerticalPadding?: number;
-
-  // ─── Typography ──────────────────────────────────────────────────────
-
-  /**
-   * Font configuration for the selector labels and captions.
-   *
-   * Uses the same `FontConfig` shape as PaymentSheet:
-   * - `family`: font family name (PostScript name on iOS, resource name on Android)
-   * - `scale`: multiplier applied to all font sizes (must be > 0)
-   *
-   * @default System font at platform-default scale
-   */
-  font?: Partial<FontConfig>;
-
-  // ─── Content ─────────────────────────────────────────────────────────
-
-  /**
-   * Controls what is displayed in each currency option's label.
-   *
-   * @default CurrencySelectorLabelContent.Automatic
-   */
-  labelContent?: CurrencySelectorLabelContent;
-}
-
 /** Represents edge insets */
 export interface EdgeInsetsConfig {
   top?: number;
